+++ src/classes/vault.ts
@@ -9,7 +9,7 @@ import {
   SystemProgram,
   TransactionInstruction,
 } from '@solana/web3.js';
-import { TOKEN_PROGRAM_ID } from '@solana/spl-token';
+import { TOKEN_PROGRAM_ID, unpackAccount } from '@solana/spl-token';
 import {
   getAssociatedTokenAddress,
   getAtasWithCreateIxnsIfMissing,
@@ -17,6 +17,7 @@ import {
   getTokenOracleData,
   KaminoReserve,
   LendingMarket,
+  MarketWithAddress,
   PubkeyHashMap,
   Reserve,
   WRAPPED_SOL_MINT,
@@ -38,7 +39,6 @@ import { VaultState } from '../idl_codegen_kamino_vault/accounts';
 import Decimal from 'decimal.js';
 import { numberToLamportsDecimal, parseTokenSymbol } from './utils';
 import { deposit } from '../idl_codegen_kamino_vault/instructions/deposit';
-import { MarketWithAddress } from './manager';
 import { withdraw } from '../idl_codegen_kamino_vault/instructions/withdraw';
 import { PROGRAM_ID } from '../idl_codegen/programId';
 import { DEFAULT_RECENT_SLOT_DURATION_MS, ReserveWithAddress } from './reserve';
@@ -81,7 +81,13 @@ export class KaminoVaultClient {
     return this._kaminoVaultProgramId;
   }
 
-  async createVault(vaultConfig: KaminoVaultConfig): Promise<[Keypair, TransactionInstruction[]]> {
+  /**
+   * This method will create a vault with a given config. The config can be changed later on, but it is recommended to set it up correctly from the start
+   * @param vaultConfig - the config object used to create a vault
+   * @returns vault - keypair, should be used to sign the transaction which creates the vault account
+   * @returns ixns - an array of instructions to create the vault
+   */
+  async createVaultIxs(vaultConfig: KaminoVaultConfig): Promise<{ vault: Keypair; ixns: TransactionInstruction[] }> {
     const vaultState = Keypair.generate();
     const size = VaultState.layout.span + 8;
 
@@ -123,10 +129,16 @@ export class KaminoVaultClient {
 
     // TODO: Add logic to update vault based on vaultConfig
 
-    return [vaultState, [createVaultIx, initVaultIx]];
+    return { vault: vaultState, ixns: [createVaultIx, initVaultIx] };
   }
 
-  async updateReserveAllocation(
+  /**
+   * This method updates the vault reserve allocation cofnig for an exiting vault reserve, or adds a new reserve to the vault if it does not exist.
+   * @param vault - vault to be updated
+   * @param reserveAllocationConfig - new reserve allocation config
+   * @returns - a list of instructions
+   */
+  async updateReserveAllocationIxs(
     vault: KaminoVault,
     reserveAllocationConfig: ReserveAllocationConfig
   ): Promise<TransactionInstruction> {
@@ -159,10 +171,17 @@ export class KaminoVaultClient {
     );
   }
 
-  async deposit(user: PublicKey, vault: KaminoVault, tokenAmount: Decimal): Promise<TransactionInstruction[]> {
+  /**
+   * This function creates instructions to deposit into a vault. It will also create ATA creation instructions for the vault shares that the user receives in return
+   * @param user - user to deposit
+   * @param vault - vault to deposit into
+   * @param tokenAmount - token amount to be deposited, in decimals (will be converted in lamports)
+   * @returns - an array of instructions to be used to be executed
+   */
+  async depositIxs(user: PublicKey, vault: KaminoVault, tokenAmount: Decimal): Promise<TransactionInstruction[]> {
     const vaultState = await vault.getState(this._connection);
 
-    const userTokenAta = await getAssociatedTokenAddress(vaultState.tokenMint, user);
+    const userTokenAta = getAssociatedTokenAddress(vaultState.tokenMint, user);
     const createAtasIxns: TransactionInstruction[] = [];
     const closeAtasIxns: TransactionInstruction[] = [];
     if (vaultState.tokenMint.equals(WRAPPED_SOL_MINT)) {
@@ -220,7 +239,15 @@ export class KaminoVaultClient {
     return [...createAtasIxns, depositIx, ...closeAtasIxns];
   }
 
-  async withdraw(
+  /**
+   * This function will return the missing ATA creation instructions, as well as one or multiple withdraw instructions, based on how many reserves it's needed to withdraw from. This might have to be split in multiple transactions
+   * @param user - user to withdraw
+   * @param vault - vault to withdraw from
+   * @param shareAmount - share amount to withdraw, in order to withdraw everything, any value > user share amount
+   * @param slot - current slot, used to estimate the interest earned in the different reserves with allocation from the vault
+   * @returns an array of instructions to be executed
+   */
+  async withdrawIxs(
     user: PublicKey,
     vault: KaminoVault,
     shareAmount: Decimal,
@@ -228,7 +255,7 @@ export class KaminoVaultClient {
   ): Promise<TransactionInstruction[]> {
     const vaultState = await vault.getState(this._connection);
 
-    const userSharesAta = await getAssociatedTokenAddress(vaultState.sharesMint, user);
+    const userSharesAta = getAssociatedTokenAddress(vaultState.sharesMint, user);
     const { atas, createAtasIxns } = await getAtasWithCreateIxnsIfMissing(
       this._connection,
       user,
@@ -237,7 +264,7 @@ export class KaminoVaultClient {
     );
     const userTokenAta = atas[0];
 
-    const tokensToWithdraw = shareAmount.div(await this.getTokensPerShare(vault, slot));
+    const tokensToWithdraw = shareAmount.div(await this.getTokensPerShareSingleVault(vault, slot));
     let tokenLeftToWithdraw = tokensToWithdraw;
 
     tokenLeftToWithdraw = tokenLeftToWithdraw.sub(new Decimal(vaultState.tokenAvailable.toString()));
@@ -305,7 +332,12 @@ export class KaminoVaultClient {
     return [...createAtasIxns, ...withdrawIxns];
   }
 
-  async investAllReserves(vault: KaminoVault): Promise<TransactionInstruction[]> {
+  /**
+   * This will trigger invest by balancing, based on weights, the reserve allocations of the vault. It can either withdraw or deposit into reserves to balance them. This is a function that should be cranked
+   * @param kaminoVault - vault to invest from
+   * @returns - an array of invest instructions for each invest action required for the vault reserves
+   */
+  async investAllReservesIxs(vault: KaminoVault): Promise<TransactionInstruction[]> {
     //TODO: Order invest ixns by - invest that removes first, then invest that adds
 
     const vaultState = await vault.getState(this._connection);
@@ -316,13 +348,19 @@ export class KaminoVaultClient {
       if (reserveState === null) {
         throw new Error(`Reserve ${reserve.toBase58()} not found`);
       }
-      investIxns.push(await this.investSingleReserve(vault, { address: reserve, state: reserveState }));
+      investIxns.push(await this.investSingleReserveIxs(vault, { address: reserve, state: reserveState }));
     }
 
     return investIxns;
   }
 
-  async investSingleReserve(vault: KaminoVault, reserve: ReserveWithAddress): Promise<TransactionInstruction> {
+  /**
+   * This will trigger invest by balancing, based on weights, the reserve allocation of the vault. It can either withdraw or deposit into the given reserve to balance it
+   * @param kaminoVault - vault to invest from
+   * @param reserve - reserve to invest into or disinvest from
+   * @returns - an array of invest instructions for each invest action required for the vault reserves
+   */
+  async investSingleReserveIxs(vault: KaminoVault, reserve: ReserveWithAddress): Promise<TransactionInstruction> {
     const vaultState = await vault.getState(this._connection);
 
     const cTokenVault = getCTokenVaultPda(reserve.address, this._kaminoVaultProgramId);
@@ -404,19 +442,73 @@ export class KaminoVaultClient {
     return withdrawIxn;
   }
 
-  async getUserSharesBalance(user: PublicKey, vault: KaminoVault): Promise<Decimal> {
+  /**
+   * This method returns the user shares balance for a given vault
+   * @param user - user to calculate the shares balance for
+   * @param vault - vault to calculate shares balance for
+   * @returns - user share balance in decimal (not lamports)
+   */
+  async getUserSharesBalanceSingleVault(user: PublicKey, vault: KaminoVault): Promise<Decimal> {
     const vaultState = await vault.getState(this._connection);
-    const userSharesAta = await getAssociatedTokenAddress(vaultState.sharesMint, user);
-    const userSharesAccount = await this._connection.getAccountInfo(userSharesAta);
-
-    if (!userSharesAccount) {
+    const userSharesAta = getAssociatedTokenAddress(vaultState.sharesMint, user);
+    const userSharesAccountInfo = await this._connection.getAccountInfo(userSharesAta);
+    if (!userSharesAccountInfo) {
       return new Decimal(0);
     }
+    const userSharesAccount = unpackAccount(userSharesAta, userSharesAccountInfo);
 
-    return new Decimal(userSharesAccount.lamports).div(new Decimal(10).pow(vaultState.sharesMintDecimals.toNumber()));
+    return new Decimal(new Decimal(userSharesAccount.amount.toString()).toNumber()).div(
+      new Decimal(10).pow(vaultState.sharesMintDecimals.toNumber())
+    );
   }
 
-  async getTokensPerShare(vault: KaminoVault, slot: number): Promise<Decimal> {
+  /**
+   * This method returns the user shares balance for all existing vaults
+   * @param user - user to calculate the shares balance for
+   * @param vaultsOverride - the kamino vaults if already fetched, in order to reduce rpc calls
+   * @returns - hash map with keyh as vault address and value as user share balance in decimal (not lamports)
+   */
+  async getUserSharesBalanceAllVaults(
+    user: PublicKey,
+    vaultsOverride?: Array<KaminoVault>
+  ): Promise<PubkeyHashMap<PublicKey, Decimal>> {
+    const vaults = vaultsOverride ? vaultsOverride : await this.getAllVaults();
+    // stores vault address for each userSharesAta
+    const vaultUserShareBalance = new PubkeyHashMap<PublicKey, Decimal>();
+    const userSharesAtaArray: PublicKey[] = [];
+    vaults.forEach((vault) => {
+      const state = vault.state;
+      if (!state) {
+        throw new Error(`Vault ${vault.address.toBase58()} not fetched`);
+      }
+      const userSharesAta = getAssociatedTokenAddress(state.sharesMint, user);
+      userSharesAtaArray.push(userSharesAta);
+    });
+    const userSharesAtaAccounts = await this._connection.getMultipleAccountsInfo(userSharesAtaArray);
+
+    userSharesAtaAccounts.forEach((userShareAtaAccount, index) => {
+      if (!userShareAtaAccount) {
+        vaultUserShareBalance.set(vaults[index].address, new Decimal(0));
+      } else {
+        vaultUserShareBalance.set(
+          vaults[index].address,
+          new Decimal(userShareAtaAccount.lamports).div(
+            new Decimal(10).pow(vaults[index].state!.sharesMintDecimals.toNumber())
+          )
+        );
+      }
+    });
+
+    return vaultUserShareBalance;
+  }
+
+  /**
+   * This method calculates the token per shar value. This will always change based on interest earned from the vault, but calculating it requires a bunch of rpc requests. Caching this for a short duration would be optimal
+   * @param vault - vault to calculate tokensPerShare for
+   * @param slot - current slot, used to estimate the interest earned in the different reserves with allocation from the vault
+   * @returns - token per share value
+   */
+  async getTokensPerShareSingleVault(vault: KaminoVault, slot: number): Promise<Decimal> {
     const vaultState = await vault.getState(this._connection);
     const reserves = await this.loadVaultReserves(vaultState);
 
@@ -445,6 +537,61 @@ export class KaminoVaultClient {
     return new Decimal(vaultState.sharesIssued.toString()).div(totalVaultLiquidityAmount);
   }
 
+  /**
+   * This method calculates the token per share value. This will always change based on interest earned from the vault, but calculating it requires a bunch of rpc requests. Caching this for a short duration would be optimal
+   * @param vault - vault to calculate tokensPerShare for
+   * @param slot - current slot, used to estimate the interest earned in the different reserves with allocation from the vault
+   * @returns - token per share value
+   */
+  async getTokensPerShareAllVaults(
+    slot: number,
+    vaultsOverride?: Array<KaminoVault>
+  ): Promise<PubkeyHashMap<PublicKey, Decimal>> {
+    const vaults = vaultsOverride ? vaultsOverride : await this.getAllVaults();
+    const vaultTokensPerShare = new PubkeyHashMap<PublicKey, Decimal>();
+    vaults.forEach(async (vault) => {
+      const vaultState = vault.state;
+      if (!vaultState) {
+        throw new Error(`Vault ${vault.address.toBase58()} not fetched`);
+      }
+      const reserves = await this.loadVaultReserves(vaultState);
+
+      const totalVaultLiquidityAmount = new Decimal(vaultState.tokenAvailable.toString());
+      vaultState.vaultAllocationStrategy.forEach((allocationStrategy) => {
+        if (!allocationStrategy.reserve.equals(PublicKey.default)) {
+          const reserve = reserves.get(allocationStrategy.reserve);
+          if (reserve === undefined) {
+            throw new Error(`Reserve ${allocationStrategy.reserve.toBase58()} not found`);
+          }
+          const reserveCollExchangeRate = reserve.getEstimatedCollateralExchangeRate(
+            slot,
+            new Fraction(reserve.state.liquidity.absoluteReferralRateSf)
+              .toDecimal()
+              .div(reserve.state.config.protocolTakeRatePct / 100)
+              .floor()
+              .toNumber()
+          );
+          const reserveAllocationLiquidityAmount = new Decimal(allocationStrategy.cTokenAllocation.toString()).div(
+            reserveCollExchangeRate
+          );
+          totalVaultLiquidityAmount.add(reserveAllocationLiquidityAmount);
+        }
+      });
+      vaultTokensPerShare.set(
+        vault.address,
+        new Decimal(vaultState.sharesIssued.toString()).div(totalVaultLiquidityAmount)
+      );
+    });
+
+    return vaultTokensPerShare;
+  }
+
+  /**
+   * This will return an unsorted hash map of all reserves that the given vault has allocations for, toghether with the amount that can be withdrawn from each of the reserves
+   * @param vault - the kamino vault to get available liquidity to withdraw for
+   * @param slot - current slot
+   * @returns an HashMap of reserves (key) with the amount available to withdraw for each (value)
+   */
   private async getReserveAllocationAvailableLiquidityToWithdraw(
     vault: KaminoVault,
     slot: number
@@ -522,7 +669,41 @@ export class KaminoVaultClient {
 
     return kaminoReserves;
   }
-}
+
+  /**
+   * Get all vaults
+   * @returns an array of all vaults
+   */
+  async getAllVaults(): Promise<KaminoVault[]> {
+    const { getProgramAccounts } = await import('../utils/rpc');
+    const filters = [
+      {
+        dataSize: VaultState.layout.span + 8,
+      },
+    ];
+
+    const [, kaminoVaults] = await Promise.all([
+      this._connection.getSlot(),
+      getProgramAccounts(this._connection, this._kaminoVaultProgramId, {
+        commitment: this._connection.commitment ?? 'processed',
+        filters,
+      }),
+    ]);
+
+    return kaminoVaults.map((kaminoVault) => {
+      if (kaminoVault.account === null) {
+        throw new Error('Invalid account');
+      }
+
+      const kaminoVaultAccount = VaultState.decode(kaminoVault.account.data);
+      if (!kaminoVaultAccount) {
+        throw Error('Could not parse obligation.');
+      }
+
+      return new KaminoVault(kaminoVault.pubkey, kaminoVaultAccount, this._kaminoVaultProgramId);
+    });
+  }
+} // KaminoVaultClient
 
 export class KaminoVault {
   readonly address: PublicKey;
@@ -548,8 +729,12 @@ export class KaminoVault {
     }
   }
 
-  async reload(connection: Connection): Promise<void> {
+  async reloadState(connection: Connection): Promise<VaultState> {
     this.state = await VaultState.fetch(connection, this.address, this.programId);
+    if (!this.state) {
+      throw new Error('Could not fetch vault');
+    }
+    return this.state;
   }
 }
