+++ src/utils/index.ts
@@ -13,3 +13,4 @@ export * from './pubkey';
 export * from './oracle';
 export * from './lookupTable';
 export * from './sendTransactionsUtils';
+export * from './managerTypes';
